Refactor main with new Auth.Service - s3i11.2
