+++ src/service/AuthenticationService.kt
@@ -10,4 +10,13 @@ class AuthenticationService {
         User(login = "aleksandra", password = "abc")
     )
 
+    fun validateLogin(login: String): Boolean {
+        return login.matches(regex = "[a-z]{1,10}".toRegex())
+    }
+
+    fun getUser(login: String): User? = users.find { it.login == login }
+
+    fun verifyPassForUser(pass: String, user: User): Boolean {
+        return user.password == pass
+    }
 }
